se corrieron errores y realizaron correcciones
