+++ NEWBIE/TODO tip-calculator-app-main/elementosHTML.js
