+++ NEWBIE/TODO tip-calculator-app-main/styless.css
@@ -21,6 +21,7 @@ html {
   font-size: 62.5%;
   font-family: "Space Mono", monospace;
 }
+
 body {
   width: 100%;
   height: 100vh;
@@ -109,6 +110,9 @@ input.into {
 }
 .tip-custom-option:focus {
   border: solid var(--Strong-cyan);
+}
+.content-tips input.event-error:focus{
+  border: solid red;
 }
 .tip-custom-option::placeholder {
   color: var(--Dark-grayish-cyan);
